// types/dto.ts
export interface BookDto {
  isbn: string;
  title: string;
  summary: string;
  publisher: Simple;
  /** @format date-time */
  publishedDate: string;
  authorList: Simple[];
  titleImage?: string;
  translator?: string;
  /** @format int32 */
  price?: number;
}

export interface Detail {
  /** @format int64 */
  isbn: string;
  title: string;
  detailUrl: string;
  summary: string;
  publishedDate: Date;
  titleImage: string;
  authorList: [];
  translator: [];
  price: number;
  status: string;
  publisher: {
    id: number;
    name: string;
  };
  contentsDtoList: [];
  eventDtoList: [];
}

export interface Simple {
  /** @format int64 */
  id: number;
  name: string;
}
// ...
export interface PageResponseBookDto {
  items: BookDto[];
  /** @format int64 */
  totalCount: number;
  /** @format int32 */
  totalPages: number;
  hasNext: boolean;
  hasPrevious: boolean;
}